+++ src/test/java/com/masonord/harmonyhound/util/LanguageUtilTest.java
@@ -3,7 +3,6 @@ package com.masonord.harmonyhound.util;
 import org.junit.jupiter.api.BeforeEach;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import org.junit.jupiter.api.Test;
-import org.springframework.cglib.core.Local;
 
 import java.util.Locale;
